// types.ts
export type ContentType = 'h1' | 'h2' | 'p' | 'list' | 'table' | 'line' | 'button' | 'spacer' | 'terminal' | 'image' | 'collage' | 'stats' | 'image-generator' | 'hero' | 'quest-log' | 'widget-selector' | 'project-list' | 'project-detail' | 'blog-news' | 'conspiracy-map' | 'event-editor' | 'resume-viewer' | 'pixel-paint' | 'display-settings' | 'system-monitor' | 'file-manager' | 'chatbot' | 'video-player' | 'browser';

export interface TableHeader {
  text: string;
  width: string;
}

export interface TagCell {
  type: 'tags';
  tags: string[];
}

export interface LinkCell {
  type: 'link';
  text: string;
  action?: string;
  url?: string;
}

export interface IconCell {
  type: 'icon';
  src: string;
  alt?: string;
}

export type TableCell = string | TagCell | LinkCell | IconCell;

export interface StatItem {
  label: string;
  value: number; // 0 to 100
  color?: string;
}

export interface QuestItem {
  title: string;
  organization: string;
  period: string;
  description: string;
  image?: string;
  tags?: string[];
  status: 'active' | 'completed';
}

export interface ContentItem {
  type: ContentType;
  text?: string;
  sub?: string; // For H1
  headers?: TableHeader[]; // For Table
  rows?: TableCell[][]; // For Table
  stats?: StatItem[]; // For Stats
  quests?: QuestItem[]; // For Quest Log
  action?: string; // For Button or internal Link
  link?: string; // For External Links
  src?: string; // For Images or Iframe URL
  secondarySrc?: string; // For Collage or Hero Character
  height?: number | string; // For Spacer or Image height
  alt?: string;
  projectId?: number; // For Project Detail
  initialPath?: string; // For File Manager
}

// --- Window Definition ---

export interface WindowHistoryItem {
    title: string;
    content: ContentItem[];
}

export interface WindowDef {
  id: string;
  title: string; // Current title displayed
  x: number;
  y: number;
  w: number;
  h: number;
  zIndex: number;
  isActive: boolean;
  isMinimized: boolean;
  isMaximized: boolean; // New: Fullscreen support
  workspace: number;
  
  // Navigation State
  history: WindowHistoryItem[];
  historyIndex: number;
}